WIP: partially refactored, working
